created new files and controller for dashboard stuff
