+++ RR.UnitOfWork/IUnitOfWork.cs
@@ -42,7 +42,6 @@ public interface IUnitOfWork
     ITerminationRepository Termination { get; }
     IEmailTemplateRepository EmailTemplate { get; }
     IEmailHistoryRepository EmailHistory { get; }
-
     Task RawSql(string sql, params NpgsqlParameter[] parameters);
     Task<string> RawSqlGet(string sql, params NpgsqlParameter[] parameters);
     Task<List<string>> GetColumnNames(string tableName);
